added camera manager and fixed character movement based on starting camera angle and position

+++ Programming3JuneGameJam/Assets/Scripts/ScriptableObjects/CharacterStats/CharacterStats.cs
@@ -6,6 +6,6 @@ using UnityEngine;
 public class CharacterStats : ScriptableObject
 {
     public float basicHP;
-    public int basicAtk;
-    public int basicSpd;
+    public float basicAtk;
+    public float basicSpd;
 }
